+++ utils/constant.ts
@@ -5,6 +5,9 @@ export const MAILS = 'mails';
 export const ABOUT_US = 'aboutUs';
 export const DOWNLOAD = 'download';
 
+export const MAILS_SLUG = '/mails';
+export const DOCS_SLUG = '/docs';
+
 export const MOBILE_WIDTH = 740;
 
 export const REVALIDATE_TIME = 60 * 60; // 1 hour to avoid github spam
